tests/requirements.txt
pylint==3.3.4
pytest==8.3.4
pytest-freezegun==0.4.2
# Currently required for pytest-freezegun, can remove once they have that fixed
setuptools==75.8.0
tox==4.24.1